Android sign in and sign up

+++ final_qualifying_work/Projects/data_provider_app/app/src/main/java/com/example/data_provider_app/dto/ErrorDto.kt
@@ -3,6 +3,6 @@ package com.example.data_provider_app.dto
 data class ErrorDto(
     val type: String,
     val title: String,
-    val status: UShort,
+    val status: Int,
     val traceId: String
 )
