+++ final_qualifying_work/Projects/data_provider_app/app/src/main/java/com/example/data_provider_app/util/ApiResult.kt
@@ -3,7 +3,7 @@ package com.example.data_provider_app.util
 sealed class ApiResult<out T> {
     data class Success<T>(val data: T?) : ApiResult<T>()
     data class Error(
-        val code: UShort,
+        val code: Int,
         val error: String
     ) : ApiResult<Nothing>()
